final refactor for  today screen to get datas
