+++ app/src/main/java/co/develhope/meteoapp/today/getWeatherIconbasedonId.kt
@@ -1,0 +1,12 @@
+package co.develhope.meteoapp.today
+
+import co.develhope.meteoapp.data.domain.WeatherIcon
+
+fun getWeatherIconbasedonId(forecastIndex: Int): WeatherIcon {
+       return  when(forecastIndex){
+            1-> WeatherIcon.SUN
+            2 -> WeatherIcon.RAIN
+            3 -> WeatherIcon.SUN_CLOUD
+            else -> WeatherIcon.NOTAVAILABLE
+        }
+    }
